conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
ca-certificates=2019.9.11=hecc5488_0
certifi=2019.9.11=py37_0
icu=64.2=he025d50_1
jpeg=9c=hfa6e2cd_1001
libpng=1.6.37=h7602738_0
openssl=1.1.1d=hfa6e2cd_0
pip=19.3.1=py37_0
pyqt=5.9.2=py37h6538335_4
python=3.7.3=h510b542_1
qt=5.9.7=h506e8af_3
setuptools=41.6.0=py37_1
sip=4.19.8=py37h6538335_0
sqlite=3.30.1=hfa6e2cd_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_0
wheel=0.33.6=py37_0
wincertstore=0.2=py37_1003
xlsxwriter=1.2.5=py_0
zlib=1.2.11=h2fa13f4_1006

